requirements.txt
# Core OCR dependencies
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.21.0

# PaddleOCR
paddlepaddle>=2.5.0
paddleocr>=3.0.0
rapidfuzz>=2.0.0

# AWS dependencies
boto3>=1.26.0
sagemaker>=2.140.0

# Data processing
pandas>=1.5.0
pyyaml>=6.0
scikit-image>=0.19.0
lmdb>=1.4.0

# Image processing and augmentation
imgaug>=0.4.0
albumentations>=1.3.0

# Machine learning
scipy>=1.9.0
matplotlib>=3.6.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
